+++ src/main/java/ch/psc/domain/storage/StorageManager.java
@@ -2,9 +2,9 @@ package ch.psc.domain.storage;
 
 import java.util.List;
 import java.util.concurrent.Future;
+import ch.psc.datasource.datastructure.Tree;
 import ch.psc.domain.file.File;
 import ch.psc.domain.storage.service.FileStorage;
-import ch.psc.persistence.datastructure.Tree;
 
 public class StorageManager {
